working on authentication

+++ Ecommerce/client/src/Pages/Home.jsx
@@ -9,13 +9,12 @@ import Filter from '../Components/Filter/Filter.jsx';
 import { Link, useNavigate } from 'react-router-dom';
 import { FaCartPlus, FaHeart, FaShoppingCart } from 'react-icons/fa';
 import { useDispatch, useSelector } from 'react-redux';
-// import { addToWishlist, removeFromWishlist } from '../Redux/Slice/WishlistSlice.js';
 import { addToCart } from '../Redux/Slice/CartSlice.js';
 import { addToWishlist, removeFromWishlist } from '../Redux/Slice/WishlistSlice.js';
 
 function Home(props) {
     const contextData = useContext(mycontext);
-    const { getAllProduct, allProduct, loading } = contextData;
+    const { getAllProduct, allProduct, loading, auth } = contextData;
 
     const [fiterProducts, setFilterProducts] = useState([]);
     const [radio, setRadio] = useState([]);
@@ -29,7 +28,7 @@ function Home(props) {
         return cartItems.some(cartItem => cartItem._id === item._id);
     };
 
-    const wishlistItems = useSelector(state => state.wishlist.wishlistItems || []); // Extracting wishlist items from Redux store
+    const wishlistItems = useSelector(state => state.wishlist.wishlistItems || []);
     const isItemInWishlist = (item) => {
         return wishlistItems.some(wishItem => wishItem._id === item._id);
     };
@@ -39,19 +38,89 @@ function Home(props) {
             navigate('/cart');
         } else {
             dispatch(addToCart(item));
-            toast.success("Product Added");
+      
+            toast.success("Product Added",
+
+            {
+                position: "top-center",
+                autoClose: 2000,
+                hideProgressBar: false,
+                closeOnClick: true,
+                pauseOnHover: true,
+                draggable: true,
+                progress: undefined,
+                theme: "dark",
+                style: {
+                    borderRadius: '10px', // Set border radius
+                    // Other CSS properties...
+                  },
+                
+            }
+
+        );
+
         }
     }
 
     const wishlistHandler = (item) => {
-        if (isItemInWishlist(item)) {
-            dispatch(removeFromWishlist(item)); // Remove item from wishlist if already in wishlist
-            toast.success("Removed Wishlist");
-       
+        if (auth.token) {
+            if (isItemInWishlist(item)) {
+                dispatch(removeFromWishlist(item));
+               
+                toast.success("Removed Wishlist",
+
+                {
+                    position: "top-center",
+                    autoClose: 2000,
+                    hideProgressBar: false,
+                    closeOnClick: true,
+                    pauseOnHover: true,
+                    draggable: true,
+                    progress: undefined,
+                    theme: "dark",
+                    style: {
+                        borderRadius: '10px', // Set border radius
+                        // Other CSS properties...
+                      },
+                    
+                }
+    
+            );
+
+
+                
+            } else {
+                dispatch(addToWishlist(item));
+
+                toast.success("Wishlisted",
+
+                    {
+                        position: "top-center",
+                        autoClose: 2000,
+                        hideProgressBar: false,
+                        closeOnClick: true,
+                        pauseOnHover: true,
+                        draggable: true,
+                        progress: undefined,
+                        theme: "dark",
+                        style: {
+                            borderRadius: '10px', // Set border radius
+                            // Other CSS properties...
+                          },
+                        
+                    }
+
+                );
+
+
+            }
         } else {
-            dispatch(addToWishlist(item)); // Add item to wishlist if not in wishlist
-            toast.success("Wishlisted");
+            navigate('/signin');
         }
+    };
+
+    const checkUserJHander = () => {
+        navigate("/signin");
     };
 
     const handleCategoryChange = (e) => {
@@ -65,7 +134,25 @@ function Home(props) {
                 radio: radio
             });
             setFilterProducts(data?.products);
-            toast.success("product filtered");
+            toast.success("Product filtered",
+
+                {
+                    position: "top-center",
+                    autoClose: 5000,
+                    hideProgressBar: false,
+                    closeOnClick: true,
+                    pauseOnHover: true,
+                    draggable: true,
+                    progress: undefined,
+                    theme: "light",
+                }
+
+            );
+
+
+
+
+
         } catch (error) {
             console.error("Error filtering products:", error);
         }
@@ -92,16 +179,16 @@ function Home(props) {
     }
 
     const imageUrls = [
-        'https://media.istockphoto.com/id/1459477634/photo/organization-shelves-with-shoes-organized-and-lined-up.jpg?s=2048x2048&w=is&k=20&c=jXgUO8USRnduIeUpja5Od0x90QPjxlqlnhdHe3t2C9M=',
-        'https://images.unsplash.com/photo-1493246507139-91e8fad9978e?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=2940&q=80',
-        'https://images.unsplash.com/photo-1518623489648-a173ef7824f3?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=2762&q=80'
-    ];
+        'https://img.freepik.com/free-photo/female-legs-sneakers-with-flowers-yellow-background_185193-109402.jpg?w=1060&t=st=1712118781~exp=1712119381~hmac=c7dc0f6fe6c0a15e95c5dd0163a11f712b586ac6d5e072f006aa285d732cd020',
+        'https://img.freepik.com/free-vector/modern-sale-banner-with-product-description_1361-1259.jpg?w=1060&t=st=1712118812~exp=1712119412~hmac=bd57f2a03130e6c872eed01f5ab94cd50f059e6985ccd7ce109846461d31ead6',
+        'https://img.freepik.com/free-vector/classic-shoes-realistic-composition_1284-65193.jpg?t=st=1712118975~exp=1712122575~hmac=51613cf48572fd389bb6f58db5ef9c6ec3e71a6a71a3d5b193e8b98f24cc23be&w=1060',
+        'https://img.freepik.com/free-photo/pack-baby-shoes-with-different-designs_1203-1856.jpg?w=1060&t=st=1712119031~exp=1712119631~hmac=70e4142946bdcce0f4ce42103f3efcc1c10502f06b798c70abac8a594ee4050e'];
 
     return (
         <div>
             <LayOut>
                 <div className="bg-black  flex flex-col">
-                    <div className='  border-2 border-yellow-500 sm:h-[32rem] h-[12rem]  rounded-xl w-full mx-auto '>
+                    <div className='sm:h-[32rem] h-[12rem]  rounded-xl w-full mx-auto '>
                         <Carousel className="rounded-xl">
                             {imageUrls.map((imageUrl, index) => (
                                 <div className="relative" key={index}>
@@ -110,7 +197,6 @@ function Home(props) {
                                         alt={`image ${index + 1}`}
                                         className="sm:h-[32rem] h-[12rem] p-2 w-full object-cover rounded-xl"
                                     />
-                                    <button className="absolute  text-3xl top-[8rem] right-[17rem] sm:top-[25rem] sm:right-[70rem] bg-white text-black font-bold  sm:text-4xl px-3 py-1 rounded-lg">Button</button>
                                 </div>
                             ))}
                         </Carousel>
@@ -127,9 +213,9 @@ function Home(props) {
                             setRadio={setRadio}
                         />
                     </div>
-                    <div className="grid grid-cols-2 gap-2 sm:grid-cols-3 md:grid-cols-3 lg:grid-cols-5">
+                    <div className="grid grid-cols-2 gap-2  sm:grid-cols-3 md:grid-cols-3 lg:grid-cols-5">
                         {allProduct.map((item, index) => (
-                            <div key={item._id} className=" sm:w-[85%] bg-white   border-2  h-[95%] max-h-[auto] m-2 p-2 border-gray-500 rounded-lg shadow relative">
+                            <div key={item._id} className="sm:w-[85%] duration-300 ease-in-out transform hover:scale-90 bg-white border-2 h-[95%] max-h-[auto] m-2 p-2 border-gray-500 rounded-lg shadow relative">
                                 <Link to={`/product/${item.slug}`}>
                                     <img className="rounded-lg item-center" src={`http://localhost:3000/api/product/product-photo/${item._id}`} alt={item.name} />
                                 </Link>
@@ -137,43 +223,68 @@ function Home(props) {
                                     <h3 className="text-lg  font-bold text-black">{item.name}</h3>
                                     <p className="mt-1 text-xs text-gray-700">${item.description.slice(0, 30)}${item.description.length > 30 ? '...' : ''}</p>
                                     <p className="mt-1 text-lg text-black line-clamp-3">${item.price}</p>
-                                    <div className="flex flex-col  sm:flex-row justify-evenly mt-3 m-1">
-                                        <button
-                                            onClick={() => handleButtonClick(item)}
-                                            className="bg-black transition border-2  border-teal-300 duration-300 ease-in-out transform hover:scale-110 hover:shadow-xl text-white font-semibold sm:py-5 py-1 px-5 rounded-lg flex items-center"
-                                        >
-                                            {isItemInCart(item) ? (
-                                                <FaShoppingCart className="mr-2" />
-                                            ) : (
-                                                <FaCartPlus className="mr-2" />
-                                            )}
-                                            <span style={{
-                                                color: isItemInCart(item) ? 'green' : 'white',
-                                                fontSize: '15px'
-                                            }}>
-                                                {isItemInCart(item) ? 'Go to Cart' : 'Add to Cart'}
-                                            </span>
-                                        </button>
-                                        <button
-                                            onClick={() => wishlistHandler(item)}
-                                            className="bg-black transition border-2 border-teal-300 duration-300 ease-in-out transform hover:scale-90 hover:shadow-xl text-white font-semibold sm:py-5 py-1 px-4 rounded-lg flex items-center"
-                                        >
-                                            <FaHeart className={`mr-1 ${isItemInWishlist(item) ? 'text-red-500' : ''}`} /> {/* Toggling red color */}
-                                            <span style={{
-                                                color: isItemInWishlist(item) ? 'red' : 'white',
-                                                fontSize: '15px'
-                                            }}>
-                                                {isItemInWishlist(item) ? 'Whislisted' : 'Wishlist'}
-                                            </span>
-                                        </button>
+                                    <div className="flex flex-col sm:flex-row justify-evenly mt-3 m-1">
+                                        {auth?.token ? (
+                                            <>
+                                                <button
+                                                    onClick={() => handleButtonClick(item)}
+                                                    className="bg-black transition border-2 border-teal-300 duration-300 ease-in-out transform hover:scale-90 hover:shadow-xl text-white font-semibold sm:py-2 sm:px-2 py-1 px-5 rounded-lg flex items-center"
+                                                >
+                                                    {isItemInCart(item) ? (
+                                                        <FaShoppingCart className="mr-2" />
+                                                    ) : (
+                                                        <FaCartPlus className="mr-2" />
+                                                    )}
+                                                    <span style={{
+                                                        color: isItemInCart(item) ? 'green' : 'white',
+                                                        fontSize: '15px'
+                                                    }}>
+                                                        {isItemInCart(item) ? 'Go to Cart' : 'Add to Cart'}
+                                                    </span>
+                                                </button>
+                                                <button
+                                                    onClick={() => wishlistHandler(item)}
+                                                    className="bg-black transition border-2 border-teal-300 duration-300 ease-in-out transform hover:scale-90 hover:shadow-xl text-white font-semibold sm:py-5 sm:px-2 py-1 px-4 rounded-lg flex items-center"
+                                                >
+                                                    <FaHeart className={`mr-1 ${isItemInWishlist(item) ? 'text-red-500' : ''}`} />
+                                                    <span style={{
+                                                        color: isItemInWishlist(item) ? 'red' : 'white',
+                                                        fontSize: '15px'
+                                                    }}>
+                                                        {isItemInWishlist(item) ? 'Whislisted' : 'Wishlist'}
+                                                    </span>
+                                                </button>
+                                            </>
+                                        ) : (
+                                            <>
+                                                <button
+                                                    onClick={checkUserJHander}
+                                                    className="bg-black transition border-2 border-teal-300 duration-300 ease-in-out transform hover:scale-90 hover:shadow-xl text-white font-semibold sm:py-2 sm:px-2 py-1 px-5 rounded-lg flex items-center"
+                                                >
+                                                    <FaShoppingCart className="mr-2" />
+                                                    <span>
+                                                        Add to Cart
+                                                    </span>
+                                                </button>
+                                                <button
+                                                    className="bg-black transition border-2 border-teal-300 duration-300 ease-in-out transform hover:scale-90 hover:shadow-xl text-white font-semibold sm:py-5 sm:px-2 py-1 px-4 rounded-lg flex items-center"
+                                                    onClick={checkUserJHander}
+                                                >
+                                                    <FaHeart className='mr-1' />
+                                                    <span>
+                                                        Wishlist
+                                                    </span>
+                                                </button>
+                                            </>
+                                        )}
                                     </div>
                                 </div>
                             </div>
                         ))}
                     </div>
                 </div>
-            </LayOut >
-        </div >
+            </LayOut>
+        </div>
     );
 }
